Ajout d'un message de confirmation à la suppr. d'un utilisateur

+++ admin.php
@@ -17,13 +17,18 @@ if ($_SESSION['admin'] == false) {
 
 $result = $blogitoDB->query("SELECT username FROM users ORDER BY username");
 ?>
-    <form method="post" action="delete_user.php">
-<?php
-foreach ($result as $row) {
 
-    $username = $row['username'];
-    echo '<a>' . $username . '</a>' . '<button type="submit" name="userToDelete" value=' . $username . '> X </button><br>';
-}
-?>
-    </form>
+<form method="post" action="delete_user.php" onsubmit="return confirm('Voulez-vous vraiment supprimer cet utilisateur ?')">
+
+    <?php
+    foreach ($result as $row) {
+
+        $username = $row['username'];
+        echo '<a>' . $username . '</a>' . '<button type="submit" name="userToDelete" value=' . $username . '> X </button><br>';
+    }
+    ?>
+
+</form>
+
+
